HEE HOO Game logic is working perfect now. but need some work on tile animation though. but game in very stable position. Can start to think about other features and can comeback later to finish off animation

+++ Game/Source/Match3/HSM.h
@@ -7,7 +7,7 @@ public:
   typedef std::map<std::pair<CBoard::State, CBoard::State>, CBoard::DoAction > TransitionMap;
   HSM();
   void Init(CBoard *board);
-  void Go(CBoard::State nextState, float delay = 0.0f);
+  void Go(CBoard::State nextState, int delay = 0);
   bool CanTransit(CBoard::State nextState);
   bool CanAcceptInput() { return m_currentState == CBoard::State::Idle || m_currentState == CBoard::State::OneItemSelected; }
   CBoard::State GetState();
@@ -19,7 +19,7 @@ private:
   CBoard::State m_currentState;
   CBoard::State m_nextState;
   float m_elapsedTime;
-  unsigned int m_targetDelay;
+  int m_targetDelay;
 
 private:
   void DoTransition();
